requirements.txt
amqp==2.1.1
billiard==3.5.0.2
celery==4.0.0
click==6.6
decorator==4.0.10
Flask==0.11.1
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
kombu==4.0.0
MarkupSafe==0.23
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
pytz==2016.7
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
Werkzeug==0.11.11
